fix: resolve merge conflict in webpack.common.js
