+++ scripts/build.js
@@ -7,7 +7,7 @@ let filesize = require('filesize');
 let gzipSize = require('gzip-size').sync;
 let rimrafSync = require('rimraf').sync;
 let webpack = require('webpack');
-let config = require('../config/webpack.config');
+let config = require('../config/webpack.prod');
 let paths = require('../config/paths');
 let recursive = require('recursive-readdir');
 let stripAnsi = require('strip-ansi');
